third_party/blink/web_tests/platform/generic/fast/webgl/context-gc-custom-properties-expected.txt

Verify that the custom properties on a WebGL rendering context object are retained across GCs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS context.customProperty is "value"
PASS extensionObject.customProperty is "value"
PASS context.customProperty is "value"
PASS extensionObject.customProperty is "value"
PASS successfullyParsed is true

TEST COMPLETE
